Restructured folder structure and made dataclasses public
