+++ src/ctek/__init__.py
@@ -1,1 +1,1 @@
-from .nanogrid_air import NanogridAir as NanogridAir
+from .nanogrid_air.nanogrid_air import NanogridAir as NanogridAir
